add routing guard

+++ src/app/shared/services/users.service.ts
@@ -4,6 +4,7 @@ import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs-compat';
 import { BlockUI, NgBlockUI } from 'ng-block-ui';
 
+import { AuthService } from './auth.service';
 import { ToastService } from './toaster.service';
 
 
@@ -16,7 +17,8 @@ export class UsersService {
     @BlockUI() blockUI: NgBlockUI;
 
     constructor(
-        private httpClient: HttpClient, 
+        private httpClient: HttpClient,
+        private auth: AuthService, 
         private toastService: ToastService) { 
     }
 
@@ -26,10 +28,10 @@ export class UsersService {
     // get all users request
     getUsers() {
         this.blockUI.start();
-        return this.httpClient.get(this.baseUrl + '/users')
+        return this.httpClient.get(this.baseUrl + '/users', { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError("Users can't be loaded!")
-                return Observable.throw(error.json().error || 'Server error')
+                this.toastService.showError(error.message || 'Server Error')
+                return Observable.throw(error.message || 'Server Error')
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -43,10 +45,10 @@ export class UsersService {
     // get single user request
     getUser(id: Number) {
         this.blockUI.start();
-        return this.httpClient.get(this.baseUrl + "/users/" + id)
+        return this.httpClient.get(this.baseUrl + "/users/" + id, { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError("User details can't be loaded!")
-                return Observable.throw(error.json().error || 'Server error')
+                this.toastService.showError(error.message || 'Server Error')
+                return Observable.throw(error.message || 'Server Error')
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -60,10 +62,10 @@ export class UsersService {
     // create single user request
     createUser(data: any) {
         this.blockUI.start();
-        return this.httpClient.post(this.baseUrl + '/users', data)
+        return this.httpClient.post(this.baseUrl + '/users', data, { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError("User can't be created!")
-                return Observable.throw(error.json().error || 'Server error')
+                this.toastService.showError(error.message || 'Server Error')
+                return Observable.throw(error.message || 'Server Error')
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -74,13 +76,29 @@ export class UsersService {
             }) as any;
     }
 
+    // updateUser(id: any, data: any, {}) {
+    //     this.blockUI.start();
+    //     return this.httpClient.delete(this.baseUrl + '/users/' + id, data, { 'headers': this.authInterceptor.headers })
+    //         .catch((error: any) => {
+    //             this.toastService.showError("User can't be deleted!")
+    //             return Observable.throw(error);
+    //         })
+    //         .finally(() => {
+    //             this.pendingRequests--;
+    //             if (this.pendingRequests <= 0) {
+    //                 // there are no more pending requests
+    //                 this.blockUI.stop();
+    //             }
+    //         }) as any;
+    // }
+
     // delete single user request
     deleteUser(id: any) {
         this.blockUI.start();
-        return this.httpClient.delete(this.baseUrl + '/users/' + id)  
+        return this.httpClient.delete(this.baseUrl + '/users/' + id, { observe: 'response', 'headers': this.auth.headers })  
             .catch((error: any) => {
-                this.toastService.showError("User can't be deleted!")
-                return Observable.throw(error);
+                this.toastService.showError(error.message || 'Server Error')
+                return Observable.throw(error.message || 'Server Error');
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -91,5 +109,4 @@ export class UsersService {
             }) as any;   
     }
 
-    
 }
